+++ pyqgisserver/handlers/owshandler.py
@@ -54,14 +54,15 @@ class OwsHandler(BaseHandler):
             if proxy_url: 
                 headers['X-Forwarded-Url']=proxy_url
 
-            def copy_headers(pat):
-                headers.update((k,v) for k,v in self.request.headers.items() if k.upper().startswith(pat))
+            def copy_headers(pats):
+                headers.update((k,v) for k,v in self.request.headers.items() if \
+                               any(map(k.upper().startswith,pats)))
 
             # Copy custom Qgis/Forwarded headers
             # see https://github.com/qgis/QGIS/pull/41333
-            copy_headers('X-QGIS-')
+            copy_headers(('X-QGIS-','X-LIZMAP-'))
 
-            if self.has_body_arguments:
+            if self.get_argument('SERVICE', default=None) and  self.has_body_arguments:
                 # Do not let qgis server handle url encoded prameters
                 data = None
                 if method == 'POST':